modify pickup location destroy action to check user confirmation before destroy; modfiy destroy integration test to check the change; add delete form in pickup time edit view

+++ app/controllers/pickup_locations_controller.rb
@@ -53,8 +53,12 @@ class PickupLocationsController < ApplicationController
 
 	def destroy
 		location = PickupLocation.find(params[:id])
-		location.destroy
-		redirect_and_flash(pickup_locations_url, :success, "Location has been removed")
+		if user_confirmed? params[:confirm_delete]
+			location.destroy
+			redirect_and_flash(pickup_locations_url, :success, "Location has been removed")
+		else
+			redirect_and_flash(edit_pickup_location_url(location), :error, "Please type I Understand to confirm delete!")
+		end
 	end
 
 	private
@@ -63,7 +67,7 @@ class PickupLocationsController < ApplicationController
 		if activating
 			location.update_attribute(:active, true)
 			redirect_and_flash(pickup_location_url(location), :success, "Location is now active")
-		elsif params[:confirm_deactive].downcase.chomp == "i understand"
+		elsif user_confirmed? params[:confirm_deactive]
 			location.update_attribute(:active, false)
 			redirect_and_flash(pickup_location_url(location), :info, "Location has been deactived")
 		else
